+++ src/materia/materia.service.ts
@@ -1,21 +1,23 @@
 import { Injectable } from '@nestjs/common';
 import { CreateMateriaDto } from './dto/create-materia.dto';
 import { UpdateMateriaDto } from './dto/update-materia.dto';
-import { PrismaService } from '../prisma/prisma.service';
+import { AcademicoPrismaService } from 'src/prisma/academico.prisma.service';
 
 @Injectable()
 export class MateriaService {
 
-  constructor(private readonly prisma: PrismaService) {}
+  constructor(private readonly prisma: AcademicoPrismaService) {}
 
   async create(createMateriaDto: CreateMateriaDto) {
-    return await this.prisma.materia.create({
+    if (!this.prisma.academico) throw new Error('BD Académico no disponible')
+      return await this.prisma.materia.create({
       data: createMateriaDto
     });
   }
 
   async findAll() {
-    return await this.prisma.materia.findMany({
+    if (!this.prisma.academico) throw new Error('BD Académico no disponible')
+      return await this.prisma.materia.findMany({
       include: {
         especialidad: {
           include: {
@@ -31,7 +33,8 @@ export class MateriaService {
   }
 
   async findOne(id: number) {
-    return await this.prisma.materia.findUnique({
+    if (!this.prisma.academico) throw new Error('BD Académico no disponible')
+      return await this.prisma.materia.findUnique({
       where: { id_materia: id },
       include: {
         especialidad: {
@@ -47,7 +50,8 @@ export class MateriaService {
   async update(id: number, updateMateriaDto: UpdateMateriaDto) {
     await this.findOne(id);
     if (updateMateriaDto.codigo) {
-      const existente = await this.prisma.materia.findFirst({
+      if (!this.prisma.academico) throw new Error('BD Académico no disponible')
+        const existente = await this.prisma.materia.findFirst({
         where: {
           codigo: updateMateriaDto.codigo,
           NOT: { id_materia: id }
@@ -57,7 +61,8 @@ export class MateriaService {
         throw new Error('Ya existe una materia con ese código.');
       }
     }
-    return await this.prisma.materia.update({
+    if (!this.prisma.academico) throw new Error('BD Académico no disponible')
+      return await this.prisma.materia.update({
       where: { id_materia: id },
       data: updateMateriaDto
     });
@@ -65,7 +70,8 @@ export class MateriaService {
 
   async remove(id: number) {
     await this.findOne(id);
-    return await this.prisma.materia.delete({
+    if (!this.prisma.academico) throw new Error('BD Académico no disponible')
+      return await this.prisma.materia.delete({
       where: { id_materia: id }
     });
   }
